requirements.txt
Flask==2.0.3
psycopg==3.0.10
psycopg2==2.8.4
psycopg2-binary==2.9.3
python-dotenv==0.20.0




